% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copydat_tsv.R
\name{copydat_tsv}
\alias{copydat_tsv}
\title{Copy data (TSV)}
\usage{
copydat_tsv(dat = get_dat(), fmt = dat2tsv)
}
\description{
Copy data, serialized in TSV (tab-separated) format, to the clipboard.
}

