% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Copy.R
\name{Copy}
\alias{Copy}
\title{Copy (a serialized version of) x to the clipboard}
\usage{
Copy(dat, fmt = dat2tsv, ...)
}
\arguments{
\item{fmt}{a function that takes (at least) an argument \code{file}}

\item{...}{further arguments to \code{writer}}

\item{object}{a serializable R object}
}
\description{
Copy (a serialized version of) x to the clipboard
}
\examples{
Copy(iris)
}

